orders/requirements.txt
django~=5.0
djangorestframework~=3.14.0
celery~=5.3.0
requests~=2.31.0
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
redis==5.2.1
drf-spectacular==0.28.0
setuptools
social-auth-app-django==5.4.2
coverage==7.6.10
django-jet-reboot==1.3.10
django-imagekit==5.0.0